Move postUtils to utils folder, refactor citation search component
