+++ app/javascript/utils/postUtils.js
